% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sonic-data.r
\name{read_hs_ascii}
\alias{read_hs_ascii}
\title{Read HS Sonic ASCII Data}
\usage{
read_hs_ascii(file_path)
}
\arguments{
\item{file_path}{A string specifying the path to the file to be read.}
}
\value{
A data.table containing the processed HS Sonic data or NULL if no valid data is found.
}
\description{
Reads and processes data from the HS Sonic ASCII format.
}
