% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht8700-functions.r
\name{read_ht8700}
\alias{read_ht8700}
\title{Read HT8700 Raw Data}
\usage{
read_ht8700(file_path)
}
\arguments{
\item{file_path}{A string specifying the path to the HT8700 file.}
}
\value{
A data.table containing the processed HT8700 data or NULL if the file is empty.
}
\description{
This function reads raw data from HT8700 files. It supports multiple file formats
and handles different structures of the data.
}
