% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sonic-data.r
\name{read_sonic}
\alias{read_sonic}
\title{Read Sonic Data}
\usage{
read_sonic(file_path)
}
\arguments{
\item{file_path}{A string specifying the path to the file to be read.}
}
\value{
A data.table containing the processed sonic data or NULL if no valid data is found.
}
\description{
Main function to read sonic data from various file formats.
}
