% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sonic-data.r
\name{read_windmaster_ascii}
\alias{read_windmaster_ascii}
\title{Read New Windmaster ASCII Data}
\usage{
read_windmaster_ascii(file_path)
}
\arguments{
\item{file_path}{A string specifying the path to the file to be read.}
}
\value{
A data.table containing the processed Windmaster data or NULL if no valid data is found.
}
\description{
Reads and processes data from the new Windmaster ASCII format.
}
